Fix corner case for non-interactive remote shells
On a remote system I had to use `exec zsh` as a dirty trick to get ZSH
running instead of Bash (because one couldn't chsh for other technical
reasons).  This snippet in ~/.profile achieved this:

    case $- in
        *i*) export SHELL=/bin/zsh; exec /bin/zsh;;
    esac

On the face of it it is smart, only running for interactive shells.  But
.bashrc (probably, or some other .*rc script) was being run first,
invoking forsay through an indirect chain of `source`s.  This caused the
following rsync error:

    protocol version mismatch -- is your shell clean?
    (see the rsync man page for an explanation)

This commit should fix that.