zsh: divest from ohmyzsh common-aliases plugin
In practice it has way more aliases than I actually use, and sometimes
they collide with real programs (which are sometimes detected and
avoided from aliasing over, but only on a case-by-case basis).

Create zsh/aliases.zsh with the aliases I actually use, which can be
expanded as time goes on.  The starting point for this set is a subset
of the aliases provided by common-aliases plugin itself.

Also add the 'aliases' plugin (not to be confused with common-aliases),
which provides the 'als' command to display an alias
cheatsheet/reference in a digestible format.